+++ Magento/Framework/Code/Generator/EntityAbstract.php
@@ -199,7 +199,7 @@ abstract class EntityAbstract
             'visibility' => 'protected',
             'docblock' => array(
                 'shortDescription' => 'Object Manager instance',
-                'tags' => array(array('name' => 'var', 'description' => '\Magento\Framework\ObjectManager'))
+                'tags' => array(array('name' => 'var', 'description' => '\Magento\Framework\ObjectManagerInterface'))
             )
         );